+++ examples/nextjs/app/search/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { SearchResultsPage } from '@caspian-explorer/script-caspian-store';
+
+export default function SearchRoute() {
+  return <SearchResultsPage />;
+}